pyproject.toml
[tool.poetry]
name = "bot-worker"
version = "0.1.0"
description = ""
authors = ["joeggg <ninjoetsu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cachetools = "^5.3.0"
pyzmq = "^25.0.0"
redis = "^4.5.1"
google-cloud-texttospeech = "^2.14.1"
requests = "^2.28.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.0.1"
pylint = "^2.16.2"
black = "^23.1.0"
pytest = "^7.2.1"
types-redis = "^4.5.1.1"
types-cachetools = "^5.3.0.0"
types-requests = "^2.28.11.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
